refactor(run): extract the await/choice/chat machinery into await-gate.ts
run.ts was two things wearing one hat: a ~600-line run lifecycle plus a ~470-line
shared library of gate/choice/chat primitives (resolveAwaitGate, requestChoices,
requestMultiSelect, drainGates, runChatPhase, runAwaitRounds and their types). The
library is what the direct-prompt path and the backlog loop reach for, not the
orchestrator, and todo-loop.ts importing requestChoices/runAwaitRounds back from run.js
was the documented run <-> todo-loop cycle. Move the library to its own module and
repoint every consumer at it; run.ts now reads as a lifecycle and imports the two
primitives it composes. No behavior change (run.ts 1070 -> 654 lines).
